+++ cvxpy/error.py
@@ -17,10 +17,12 @@ You should have received a copy of the GNU General Public License
 along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 """
 
+
 class SolverError(Exception):
     """Error thrown when a solver encounters an error solving a problem.
     """
     pass
+
 
 class DCPError(Exception):
     """Error thrown for DCP violations.