+++ cvxpy/tests/test_matrices.py
@@ -30,6 +30,8 @@ import cvxopt
 import scipy
 import scipy.sparse as sp
 import unittest
+import sys
+PY35 = sys.version_info >= (3,5)
 
 class TestMatrices(unittest.TestCase):
     """ Unit tests for testing different forms of matrices as constants. """
@@ -67,6 +69,8 @@ class TestMatrices(unittest.TestCase):
         # Matrix
         A = numpy.arange(8).reshape((4,2))
         self.assertExpression(A*self.x, (4,1))
+        if PY35:
+            self.assertExpression(self.x.__rmatmul__(A), (4,1))
         # PSD inequalities.
         A = numpy.ones((2,2))
         self.assertExpression(A << self.A, (2,2))
@@ -88,6 +92,8 @@ class TestMatrices(unittest.TestCase):
         A = numpy.matrix( numpy.arange(8).reshape((4,2)) )
         self.assertExpression(A*self.x, (4,1))
         self.assertExpression( (A.T*A) * self.x, (2,1))
+        if PY35:
+            self.assertExpression(self.x.__rmatmul__(A), (4,1))
         # PSD inequalities.
         A = numpy.matrix(numpy.ones((2,2)))
         self.assertExpression(A << self.A, (2,2))
@@ -161,6 +167,8 @@ class TestMatrices(unittest.TestCase):
         self.assertExpression(B * var, (4, 2))
         self.assertExpression(var - A, (4, 2))
         self.assertExpression(A - A - var, (4, 2))
+        if PY35:
+            self.assertExpression(var.__rmatmul__(B), (4,2))
         # self.assertExpression(var <= A, (4, 2))
         # self.assertExpression(A <= var, (4, 2))
         # self.assertExpression(var == A, (4, 2))